fix(config-ui): pipeline historical cannot auto refresh
