+++ config-ui/src/pages/blueprint/detail/components/index.ts
@@ -20,4 +20,3 @@ export * from './update-name-dialog';
 export * from './update-policy-dialog';
 export * from './update-scope-dialog';
 export * from './update-transformation-dialog';
-export * from './pipeline-list';
