+++ config-ui/src/pages/blueprint/detail/panel/status.tsx
@@ -21,17 +21,14 @@ import { Button, Switch, Intent } from '@blueprintjs/core';
 import dayjs from 'dayjs';
 
 import { getCron } from '@/config';
-import type { PipelineType } from '@/pages';
-import { PipelineInfo } from '@/pages';
+import { PipelineInfo, PipelineHistorical } from '@/pages';
 
 import type { BlueprintType } from '../../types';
 
-import { PipelineList } from '../components';
 import * as S from '../styled';
 
 interface Props {
   blueprint: BlueprintType;
-  pipelines: PipelineType[];
   pipelineId?: ID;
   operating: boolean;
   onRun: () => void;
@@ -39,7 +36,7 @@ interface Props {
   onDelete: () => void;
 }
 
-export const Status = ({ blueprint, pipelines, pipelineId, operating, onRun, onUpdate, onDelete }: Props) => {
+export const Status = ({ blueprint, pipelineId, operating, onRun, onUpdate, onDelete }: Props) => {
   const cron = useMemo(() => getCron(blueprint.isManual, blueprint.cronConfig), [blueprint]);
 
   const handleRunNow = () => onRun();
@@ -88,7 +85,7 @@ export const Status = ({ blueprint, pipelines, pipelineId, operating, onRun, onU
       </div>
       <div className="block">
         <h3>Historical Pipelines</h3>
-        <PipelineList pipelines={pipelines} />
+        <PipelineHistorical blueprintId={blueprint.id} />
       </div>
     </S.StatusPanel>
   );
