+++ config-ui/src/pages/pipeline/components/historical/styled.ts
@@ -16,6 +16,31 @@
  *
  */
 
-import { request } from '@/utils';
+import { Colors } from '@blueprintjs/core';
+import styled from 'styled-components';
 
-export const getPipelineLog = (id: ID) => request(`/pipelines/${id}/logging.tar.gz`);
+export const StatusColumn = styled.div`
+  display: flex;
+  align-items: center;
+
+  .bp4-icon {
+    margin-right: 4px;
+  }
+
+  &.ready,
+  &.cancel {
+    color: #94959f;
+  }
+
+  &.loading {
+    color: #7497f7;
+  }
+
+  &.success {
+    color: ${Colors.GREEN3};
+  }
+
+  &.error {
+    color: ${Colors.RED3};
+  }
+`;
